Updated the logic in generate_normalPlayIndexFile() method in ifs_GenTrickPlay.c to include correct 'frame_size'(field 4) values in the normal play index files. Resolved issue in ifs_H264_parse.c where an extra 1 second is being added to the time entries (ifsHandle->entry.when).

+++ ifs/ifs_GenTrickPlay.h
@@ -113,7 +113,7 @@ typedef struct _refIframeEntery
 
 typedef struct _refVobuEntry
 {
-    unsigned long  index;
+    unsigned long long  index;
     unsigned long  uIframeNum;
     IfsIndexEntry    entry;
     unsigned        pktCount;
